requirements.txt
numpy
scipy
h5py
py2neo
matplotlib
networkx
multiprocess
sklearn
scikit-bio
biom-format
wxPython
psutil